added new sftp option delete after upload. Added mp3 Manipulation

+++ ttd_helper.py
@@ -80,10 +80,15 @@ if args.tone_name is not None and args.mp3 is not None:
         config.token = zello.zello_create_token(config)
         opus_file = zello.zello_convert(config.ttd_audio_path, args.mp3.replace("./audio/", "").replace(".mp3", ""))
         zello.ZelloSend(config, opus_file).zello_init_upload()
-        if os.path.exists(opus_file):
-            os.remove(opus_file)
+        if config.zello_settings["delete_after_stream"] == 1:
+            if os.path.exists(opus_file):
+                os.remove(opus_file)
     else:
         print("Not Sending To Zello")
+
+    if config.sftp_settings["enabled"] == 1:
+        if config.sftp_settings["delete_after_upload"] == 1:
+            os.remove(mp3_path)
 
     if config.local_cleanup_settings["enabled"] == 1:
         print("Cleaning Local Files")
